+++ plugin/plugin/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.4.x
 
 ### 3.4.2.14-29-29.1
+ - Add `net.twisterrob.java-library` plugin
 
 ### 3.4.2.14-29-29.0
  - Gradle 5.6.4